+++ src/__tests__/clues.test.ts
@@ -1,4 +1,4 @@
-import { clues } from "../src/clues";
+import { clues } from "../clues";
 import TinyUrl from "tinyurl";
 import GithubSlugger from "github-slugger";
